requirements.txt
clean-fid==0.1.35
datasets==4.2.0
diffusers==0.35.2
image-reward==1.5
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
Pillow==11.3.0
PyYAML==6.0.3
scipy==1.15.3
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.8.2
torchvision==0.21.0
protobuf==6.32.1
sentencepiece==0.2.1
transformers==4.49.0